pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ia2"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pycairo",
    "pygame-ce",
    "numpy",
    "scipy",
]

[project.optional-dependencies]
opengl = ["moderngl"]
dev = ["pytest", "pytest-cov", "ruff"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
